refactor: keep shortcut observer at legacy boundary
